Add configuration for publishing-api.
This is the first to use environment variables, rather than a redis.yml which
is overwritten on deploy, to specify the Redis config. So the rack app needs to
be run under govuk_setenv so it can pick up the vars already created for the
actual publishing-api app by Puppet.