Add monitoring to Content Performance Manager
ontent Performance Manager is using Sidekiq to run jobs in the
background; as a result, we would like to monitor it within GOV.UK
infrastructure.

The port 3207 is not being listed in the Profile so I am making the
assumption that there won’t be conflicts.

See: https://github.gds/gds/development/pull/385,
https://github.gds/gds/development/pull/384