Use govuk_sidekiq rather than sidekiq directly
GOV.UK apps use govuk_sidekiq as an intermediary way to install sidekiq
which has version constraints, by having this tool also use
govuk_sidekiq those some restraints can be shared and maintained.

This has been done to avoid a situation where sidekiq-monitoring is on a
different major version of Sidekiq to the rest of GOV.UK apps.

This is not an ideal measure, this pseudo application is very delicate
in that it just hopes it's compatible with versions apps are using. It'd
be better if each app ran their own web interface.